Restructure session package
