+++ .changeset/231d07e0/changes.json
@@ -1,0 +1,1 @@
+{ "releases": [{ "name": "@keystone-alpha/session", "type": "patch" }], "dependents": [] }
