+++ .changeset/231d07e0/changes.md
@@ -1,0 +1,1 @@
+- Restructure internal code
